backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
pydantic==2.5.6
pydantic-settings==2.2.1
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
pyannote.audio==3.3.2
huggingface-hub==0.20.3
python-dotenv==1.0.0